% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ball_distribution.R
\name{ball_distribution}
\alias{ball_distribution}
\title{Simulate from a Ball Distribution}
\usage{
ball_distribution(n, p)
}
\arguments{
\item{n}{number of samples}

\item{p}{ball dimension}
}
\value{
\code{n} by \code{p} matrix with one sample in each row.
}
\description{
Produces samples from the uniform distribution on the p-ball.
}
\details{
The ball distribution is the distribution of Z where
\deqn{Z=UX,}
where U is the inverse p-th root of a uniform distribution on \eqn{[0,1]} and X is uniform on the (p-1)-sphere which is generated from \code{\link{sphere_distribution}}.
}
\examples{
set.seed(1)
u = ball_distribution(1000,2)
plot(u)
}
\seealso{
\code{\link{sphere_distribution}}
}
\author{
Leo Meurice
}
\keyword{MKdepth}
