% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banana_distribution.R
\name{banana_distribution}
\alias{banana_distribution}
\title{Simulate from a Banana Distribution}
\usage{
banana_distribution(n, height = 1, thick = 1)
}
\arguments{
\item{n}{number of samples.}

\item{height}{height of the banana. By default \code{height=1}}

\item{thick}{thickness of the banana. By default \code{thick=1}}
}
\value{
\code{n} by \code{2} matrix with one sample in each row.
}
\description{
Produces samples from a banana distribution in two dimensions.
}
\details{
The banana distribution is the distribution of the vector \deqn{(X + R cos \phi,height*X^2 + R sin \phi),}
where X is uniform on \eqn{[-1, 1]}, \eqn{\phi} is uniform on \eqn{[0, 2\pi]}, Z is uniform on \eqn{[0, thick]}, X, Z and \eqn{\phi}
are independent, and R = 0.2Z(1 + (1 - |X|)/2).
}
\examples{
set.seed(1)
u = banana_distribution(1000)
plot(u)

set.seed(1)
u = banana_distribution(1000,4,.5)
plot(u)
}
\references{
\insertRef{chernozhukov2017monge}{MKdepth}
}
\author{
Leo Meurice
}
\keyword{MKdepth}
