% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{MKdepth.package}
\alias{MKdepth.package}
\title{MKdepth Package}
\description{
This package contains functions allowing to calculate the depth of Monge-Kantorovich
and allowing to perform other similar tasks like drawing depth contours.
}
\seealso{
\code{\link{ball_distribution}}
\code{\link{banana_distribution}}
\code{\link{MKdepth}}
\code{\link{MKdepth_contours}}
\code{\link{MKdepth_iterative}}
\code{\link{MKdepth_map}}
\code{\link{MKdepth_median}}
\code{\link{MKdepth_specific}}
\code{\link{sphere_distribution}}
}
\author{
Leo Meurice \email{leomeurice@hotmail.com}
}
