% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MKdepth.R
\name{MKdepth}
\alias{MKdepth}
\title{The Monge-Kantorovich Depth}
\usage{
MKdepth(
  data,
  trs_method = "assignment",
  depth_method = "Tukey",
  approx = FALSE,
  prec = 3
)
}
\arguments{
\item{data}{\code{n} by \code{p} matrix of data where each row contains a p-variate point, w.r.t which the Monge-Kantorovich depth
is to be calculated.}

\item{trs_method}{Character sting which determines the transport algorithm used. By default, \code{trs_method="assignment"}. See details below.}

\item{depth_method}{Character string which determines the depth function used for the reference distribution. Method can be "Tukey" (the default), "Liu" or "Oja".}

\item{approx}{Logical. If dimension is greater than or equal to 3, the Tukey depth can be approximate . Useful when sample size is large.}

\item{prec}{Integer indicating the number of decimal places to be used for depth values. By default, \code{prec=3}.}
}
\value{
Numerical vector of depths, one for each row in \code{data}.
}
\description{
Computes the Monge-Kantorovich depth of each observation from a data set.
}
\details{
The Monge-Kantorovich depth is constructed as explained in \insertCite{chernozhukov2017monge}{MKdepth}.

In order to calculate the Monge-Kantorovich depth, we need to calculate the Tukey depth of a specific distribution.
For this, we use the function \code{\link[depth]{depth}} of the package \link{depth}. For more details on the depth method, please see \link[depth]{depth} details.

The parameter \code{trs_method} is the transport method which is used for the Monge-Kantoroch depth.
The default method are: \code{"assignement"}, comes from the \code{adagio} package (See References). The other methods which can be used,
come from the \code{transport} package and are the same as for the \code{\link[transport]{transport}} function (See References).
}
\examples{
set.seed(1)
data=ball_distribution(1000,2)
depth=MKdepth(data)

set.seed(1)
data=matrix(rnorm(2000),1000)
depth=MKdepth(data,trs_method="auction")
}
\references{
\insertRef{depth}{MKdepth}

\insertRef{borchers2012package}{MKdepth}

\insertRef{schuhmacher2019package}{MKdepth}

\insertRef{chernozhukov2017monge}{MKdepth}
}
\author{
Leo Meurice
}
\keyword{MKdepth}
