% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MKdepth_contours.R
\name{MKdepth_contours}
\alias{MKdepth_contours}
\title{Monge-Kantorovich Depth Contours}
\usage{
MKdepth_contours(
  data,
  depth,
  alpha,
  contours = 3,
  col = topo.colors(contours),
  legend = F,
  plot.new = T
)
}
\arguments{
\item{data}{\code{n} by \code{2} matrix of data where each row contains a two dimensional point, w.r.t which the Monge-Kantorovich depth
is calculated. \code{n} must be larger than 20.}

\item{depth}{Numerical vector of depths, one for each row in \code{data}.}

\item{alpha}{Value of \eqn{\alpha}. This parameter is defined in the function \code{\link[alphahull]{ahull}}
from the \code{\link{alphahull-package}}. (See Details below.)}

\item{contours}{Integer. Contours number to be drawn. By default, \code{contours=3}.}

\item{col}{Color, used to draw contours. The length must be equal to \code{contours}.}

\item{legend}{Logical; if \code{TRUE}, the legend is displayed. By default, \code{legend=F}.}

\item{plot.new}{Logical; if \code{TRUE}, a new plot is displayed and the data points and contours are drawn.}
}
\description{
Builds the Monge-Kantorovich depth contours for 2-dimensional data.
}
\details{
The depth contours are drawn using the \code{\link[alphahull]{ahull}} function from the \code{\link{alphahull-package}}.
The \code{alpha} parameter must be adjusted to conform to the shape of the non-convex depth contours.
}
\examples{
set.seed(1)
data=ball_distribution(100,2)
depth=MKdepth_iterative(data,iter=100)
MKdepth_contours(data,depth$MKdepth,alpha=1)

set.seed(1)
data=matrix(rnorm(2000),1000)
depth=MKdepth(data)
plot(data,pch='.')
MKdepth_contours(data,depth,contours=5,alpha=2,plot.new=FALSE)

}
\references{
\insertRef{alphahull}{MKdepth}
}
\author{
Leo Meurice
}
\keyword{MKdepth}
