% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MKdepth_iterative.R
\name{MKdepth_iterative}
\alias{MKdepth_iterative}
\title{The Monge-Kantorovich Depth Iterative Function}
\usage{
MKdepth_iterative(
  data,
  trs_method = "assignment",
  depth_method = "Tukey",
  approx = FALSE,
  iter = 1,
  eps = NULL,
  prec = 3
)
}
\arguments{
\item{data}{\code{n} by \code{p} matrix of data where each row contains a p-variate point, w.r.t which the Monge-Kantorovich depth
is to be calculated.}

\item{trs_method}{Character sting which determines the transport algorithm used. By default, \code{trs_method="assignment"}. See details below.}

\item{depth_method}{Character string which determines the depth function used for the reference distribution. Method can be "Tukey" (the default), "Liu" or "Oja".}

\item{approx}{Logical. If dimension is greater than or equal to 3, the Tukey depth can be approximate . Useful when sample size is large.}

\item{iter}{Positive integer. Iterations number, i.e , the calls number of the function \code{\link{MKdepth}}.
If \code{eps} is specified, \code{iter} is not considered. By default \code{iter=1}.}

\item{eps}{Positive double. Maximum expected difference between
the depth values of the penultimate iteration and the depth values of the last iteration.
If \code{eps=NULL}, this is \code{iter}, the iterations number, which is considered. By default \code{eps=NULL}.}

\item{prec}{Integer indicating the number of decimal places to be used for depth values. By default, \code{prec=3}.}
}
\value{
List containing a numerical vector of depths (\code{$MKdepth}), one for each row in \code{data}, the iterations number (\code{$iteration})
and the execution time (\code{$time}).
}
\description{
Calculates the Monge-Kantorovich depth
using several iterations in order to obtain  more reliable values.
The function \code{\link[MKdepth]{MKdepth}} is computed several times
to return the median of the different results.
}
\details{
Due to the random generate of a ball-distribution, the computed depth values by the
\code{\link{MKdepth}} function may not be precisely exact. In order to counter this, the function \code{\link{MKdepth}} is computed
several times and the median of all results is selected. To limit the number of iterations, the parameters
\code{iter} and \code{eps} are used.
}
\examples{
data=ball_distribution(100,2)
depth=MKdepth_iterative(data,eps=0.01)

data=matrix(rnorm(2000),1000)
depth=MKdepth_iterative(data,iter=10)
}
\author{
Leo Meurice
}
\keyword{MKdepth}
