% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MKdepth_map.R
\name{MKdepth_map}
\alias{MKdepth_map}
\title{Level plot for Monge-Kantorovich Depth}
\usage{
MKdepth_map(
  data,
  depth,
  alpha,
  by,
  col = heat.colors,
  main = "",
  xlab = "",
  ylab = ""
)
}
\arguments{
\item{data}{\code{n} by \code{2} matrix of data where each row contains a two dimensional point, w.r.t which the Monge-Kantorovich depth
is calculated.}

\item{depth}{Numerical vector of depths, one for each row in \code{data}.}

\item{alpha}{Value of \eqn{\alpha}. This parameter is defined in the function \code{\link[alphahull]{ahull}}
from the \code{\link{alphahull-package}}. (See Details below.)}

\item{by}{Interpolation step.}

\item{col}{color, used to draw level plot.}

\item{main}{an overall title for the plot (optional).}

\item{xlab}{a title for the x axis (optional).}

\item{ylab}{a title for the y axis (optional).}
}
\description{
Produces a contour plot with the areas between the contours
filled in solid color. A key showing how the colors map to depth values
is shown to the right of the plot. Construct from \code{\link[graphics]{filled.contour}} fonction.
}
\examples{
set.seed(1)
data=ball_distribution(1000,2)
depth=MKdepth(data)
MKdepth_map(data,depth,alpha=.5,by=.1)

set.seed(1)
data=matrix(rnorm(2000),1000)
depth=MKdepth(data)
MKdepth_map(data,depth,alpha=2,by=.1)

}
\references{
\insertRef{alphahull}{MKdepth}
}
\author{
Leo Meurice
}
\keyword{MKdepth}
