% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MKdepth_median.R
\name{MKdepth_median}
\alias{MKdepth_median}
\title{Monge-Kantorovich Median}
\usage{
MKdepth_median(data, depth)
}
\arguments{
\item{data}{\code{n} by \code{p} matrix of data where each row contains a p-variate point, w.r.t which the Monge-Kantorovich depth
is calculated.}

\item{depth}{Numerical vector of depths, one for each row in \code{data}.}
}
\value{
List containing the Monge-Kantorovich median coordinates(\code{$coordinates}) and its depth value(\code{$depth}).
}
\description{
Computes the Monge-Kantorovich Median of a multivariate data set.
}
\examples{
set.seed(1)
data=ball_distribution(1000,2)
depth=MKdepth(data)
median=MKdepth_median(data,depth)
plot(data,pch='.')
points(median$coordinates[1],median$coordinates[2],pch=19,col='red')

set.seed(1)
data=matrix(rnorm(2000),1000)
depth=MKdepth(data)
median=MKdepth_median(data,depth)
plot(data,pch='.')
MKdepth_contours(data,depth,contours=5,alpha=2,plot.new=FALSE)
points(median$coordinates[1],median$coordinates[2],pch=19,col='red')

}
\author{
Leo Meurice
}
\keyword{MKdepth}
