% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MKdepth_specific.R
\name{MKdepth_specific}
\alias{MKdepth_specific}
\title{The Monge-Kantorovich Depth of specific observations compared a data set}
\usage{
MKdepth_specific(
  x,
  data,
  trs_method = "assignment",
  depth_method = "Tukey",
  approx = FALSE,
  iter = 1,
  eps = NULL,
  prec = 3
)
}
\arguments{
\item{x}{\code{n} by \code{p} matrix of data  whose Monge-Kantorovich depth is to be calculated. Each row contains a p-variate point.
Should have the same dimension as \code{data}.}

\item{data}{\code{n} by \code{p} matrix of data where each row contains a p-variate point, w.r.t which the Monge-Kantorovich depth
is to be calculated.}

\item{trs_method}{Character sting which determines the transport algorithm used. By default, \code{trs_method="assignment"}. See details below.}

\item{depth_method}{Character string which determines the depth function used for the reference distribution. Method can be "Tukey" (the default), "Liu" or "Oja".}

\item{approx}{Logical. If dimension is greater than or equal to 3, the Tukey depth can be approximate . Useful when sample size is large.}

\item{iter}{Positive integer. Iterations number, i.e , the calls number of the function \code{\link{MKdepth}}.
If \code{eps} is specified, \code{iter} is not considered. By default \code{iter=1}.}

\item{eps}{Positive double. Maximum expected difference between
the depth values of the penultimate iteration and the depth values of the last iteration.
If \code{eps=NULL}, this is \code{iter}, the iterations number, which is considered. By default \code{eps=NULL}.}

\item{prec}{Integer indicating the number of decimal places to be used for depth values. By default, \code{prec=3}.}
}
\value{
Numerical vector of depths, one for each row in \code{x}.
}
\description{
Calculates the Monge-Kantorovich depth of each specific observation of a data set.
}
\details{
Call the function \code{\link{MKdepth_iterative}}.
}
\examples{
set.seed(1)
data=ball_distribution(100,2)
x=rbind(c(-0.2,0.1),c(-.25,-.5),c(.5,0))
depth=MKdepth_specific(x,data,iter=20)
plot(data,pch=20)
points(x,pch=19,col='red')
text(x[,1], x[,2], depth,cex=1, pos=3,col="red")
}
\seealso{
\code{\link{MKdepth_iterative}}
}
\author{
Leo Meurice
}
\keyword{MKdepth}
