% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere_distribution.R
\name{sphere_distribution}
\alias{sphere_distribution}
\title{Simulate from a Sphere Distribution}
\usage{
sphere_distribution(n, p)
}
\arguments{
\item{n}{number of samples}

\item{p}{sphere dimension}
}
\value{
\code{n} by \code{p} matrix with one sample in each row.
}
\description{
Produces samples from the uniform distribution on the p-sphere.
}
\details{
The sphere distribution is the distribution of Y where
\deqn{Y = \frac{X}{\Vert X \Vert},}
where X follows a multinomial distribution.
}
\examples{
u = sphere_distribution(100,2)
plot(u)
}
\author{
Leo Meurice
}
\keyword{MKdepth}
